+++ SignalServiceKit/Cryptography/Cryptography.m
@@ -100,7 +100,7 @@ const NSUInteger kAES256_KeyByteLength = 32;
 
 + (NSData *)generateRandomBytes:(NSUInteger)numberBytes
 {
-    return [Randomness generateRandomBytes:(int)numberBytes];
+    return [Randomness generateRandomBytes:numberBytes];
 }
 
 + (uint64_t)randomUInt64
